move _subtitute into new file
