+++ tests/test_web_app_steps.py
@@ -19,7 +19,7 @@ from gauge_web_app_steps.web_app_steps import (
     execute_async_script, execute_async_script_on_element, execute_async_script_on_element_save_result, execute_async_script_save_result,
     execute_script, execute_script_on_element, execute_script_on_element_save_result, execute_script_save_result,
     reset_timeout, save_placeholder, save_window_handles, save_window_title, set_timeout, switch_to_frame,
-    _substitute, timeout_key, wait_for_window
+    timeout_key, wait_for_window
 )
 
 
@@ -138,63 +138,6 @@ class TestWebAppSteps(unittest.TestCase):
     def test_set_timeout_error(self):
         self.assertRaises(AssertionError, lambda: set_timeout("id"))
 
-    def test_substitute_replace(self):
-        tableflip = "(ノಠ益ಠ)ノ彡┻━┻"
-        os.environ["placeholder"] = tableflip
-        result = _substitute("${placeholder}")
-        self.assertEqual(tableflip, result)
-
-    def test_substitute_with_full_math_expression(self):
-        result = _substitute("#{1 + 1}")
-        self.assertEqual("2", result)
-
-    def test_substitute_with_inner_math_expression(self):
-        result = _substitute("\(^#{0 + 0}^)/")
-        self.assertEqual("\(^0^)/", result)
-
-    def test_substitute_with_two_math_expressions(self):
-        result = _substitute("#{0 + 0}#{1 + 1}")
-        self.assertEqual("02", result)
-
-    def test_substitute_complex(self):
-        os.environ["a"] = "1"
-        result = _substitute("(${a} + 1) ** 2 = #{($a + 1) ** 2}")
-        self.assertEqual("(1 + 1) ** 2 = 4", result)
-
-    def test_substitute_fails(self):
-        result = _substitute("}#{1 + 1")
-        self.assertEqual("}#{1 + 1", result)
-
-    def test_substitute_with_bracket_diversions(self):
-        result = _substitute("}#{0 + 0}#{")
-        self.assertEqual("}0#{", result)
-
-    def test_substitute_without_pipe_operator(self):
-        placeholder1 = "lala"
-        placeholder2 = "baba"
-        os.environ["placeholder1"] = placeholder1
-        data_store.scenario["placeholder2"] = placeholder2
-        result = _substitute("This is ${placeholder1}/url and ${placeholder2}")
-        self.assertEqual("This is lala/url and baba", result)
-
-    def test_substitute_with_uuid_expression(self):
-        result = _substitute("!{uuid}")
-        self.assertRegex(result, "^[0-9a-f]{8}.[0-9a-f]{4}.[0-9a-f]{4}.[0-9a-f]{4}.[0-9a-f]{12}$")
-
-    def test_substitute_with_time_expression(self):
-        result = _substitute("!{time}")
-        self.assertTrue(self._datetime_valid(result))
-
-    def test_substitute_with_time_and_format_expression(self):
-        result = _substitute("!{time:%Y}")
-        self.assertTrue(re.fullmatch("[0-9]{4}", result) is not None)
-
-    def test_substitute_raises_with_invalid_time(self):
-        self.assertRaises(ValueError, lambda: _substitute("!{time-ly}"))
-
-    def test_substitute_raises_with_invalid_expression(self):
-        self.assertRaises(ValueError, lambda: _substitute("!{nonexistent}"))
-
     def test_switch_to_frame_by_index(self):
         self.app_context.driver.find_elements.return_value=[]
         self.assertRaises(AssertionError, lambda: switch_to_frame("1"))
@@ -260,12 +203,6 @@ class TestWebAppSteps(unittest.TestCase):
         ])
         self.assertEqual("result", data_store.scenario.get("placeholder"))
 
-    def _datetime_valid(self, dt_str: str) -> bool:
-        try:
-            datetime.fromisoformat(dt_str)
-        except ValueError:
-            return False
-        return True
 
 if __name__ == '__main__':
     unittest.main()
